Exception handling sort of works now - Added a bunch of macros for exceptions - When an unhandled exception occurs, a error is printed out to stdout - Lots of work needs to be done to ensure that no memory leaks occur when an exception is raised, this needs to be done on a case by case basis - There a lot of places where exceptions need to be raised <--- TODO

+++ rtlib/builtinfuncs.h
@@ -1,5 +1,5 @@
 #pragma once
-
+#include <stddef.h>
 typedef struct RtObject RtObject;
 
 /**
@@ -12,7 +12,7 @@ typedef struct BuiltinFunc
 
     RtObject *(*builtin_func)(RtObject **, int);
 
-    int arg_count; // if its -1, then functions takes any amount of arguments
+    size_t arg_count; // if its -1, then functions takes any amount of arguments
 
 } BuiltinFunc;
